+++ tests/test_MSRawData.py
@@ -13,11 +13,17 @@ WIDETABLEFORM_TRANSPOSE_RESULTS_FILENAME = os.path.join(os.path.dirname(__file__
 WIDETABLEFORM_ISO_8859_1_ENCODE_FILENAME = os.path.join(os.path.dirname(__file__),"testdata", 'ISO_8859_1_Encoding.csv')
 WIDETABLEFORM_ISO_8859_1_ENCODE_RESULTS_FILENAME = os.path.join(os.path.dirname(__file__),"testdata", 'ISO_8859_1_Encoding_Results.xlsx')
 
+WIDETABLEFORM_QUALIFIER_FILENAME = os.path.join(os.path.dirname(__file__),"testdata", 'WideTableForm_Qualifier.csv')
+WIDETABLEFORM_QUALIFIER_RESULTS_FILENAME = os.path.join(os.path.dirname(__file__),"testdata", 'WideTableForm_Qualifier_Results.xlsx')
+
 LARGE_WIDETABLEFORM_FILENAME = os.path.join(os.path.dirname(__file__),"testdata", 'LargeTestData.csv')
 LARGE_WIDETABLEFORM_RESULTS_FILENAME = os.path.join(os.path.dirname(__file__),"testdata", 'LargeTestData_Results.xlsx')
 
 COMPOUNDTABLEFORM_FILENAME = os.path.join(os.path.dirname(__file__),"testdata", 'CompoundTableForm.csv')
 COMPOUNDTABLEFORM_RESULTS_FILENAME = os.path.join(os.path.dirname(__file__),"testdata", 'CompoundTableForm_Results.xlsx')
+
+COMPOUNDTABLEFORM_QUALIFIER_FILENAME = os.path.join(os.path.dirname(__file__),"testdata", 'CompoundTableForm_Qualifier.csv')
+COMPOUNDTABLEFORM_QUALIFIER_RESULTS_FILENAME = os.path.join(os.path.dirname(__file__),"testdata", 'CompoundTableForm_Qualifier_Results.xlsx')
 
 SCIEX_FILENAME = os.path.join(os.path.dirname(__file__),"testdata", 'SciExTestData.txt')
 SCIEX_RESULTS_FILENAME = os.path.join(os.path.dirname(__file__),"testdata", 'SciExTestData_Results.xlsx')
@@ -27,15 +33,19 @@ class Agilent_Test(unittest.TestCase):
     def setUp(self):
         self.WideData = AgilentMSRawData(WIDETABLEFORM_FILENAME,ingui=True)
         self.ISO_EncodedData = AgilentMSRawData(WIDETABLEFORM_ISO_8859_1_ENCODE_FILENAME,ingui=True)
+        self.WideData_Qualifier = AgilentMSRawData(WIDETABLEFORM_QUALIFIER_FILENAME,ingui=True)
         self.LargeWideData = AgilentMSRawData(LARGE_WIDETABLEFORM_FILENAME,ingui=True)
         self.CompoundData = AgilentMSRawData(COMPOUNDTABLEFORM_FILENAME,ingui=True)
+        self.CompoundData_Qualifier = AgilentMSRawData(COMPOUNDTABLEFORM_QUALIFIER_FILENAME,ingui=True)
         self.SciexData = SciexMSRawData(SCIEX_FILENAME,ingui=True)
 
         self.WideDataResults = openpyxl.load_workbook(WIDETABLEFORM_RESULTS_FILENAME)
         self.WideDataTransposeResults = openpyxl.load_workbook(WIDETABLEFORM_TRANSPOSE_RESULTS_FILENAME)
         self.ISO_EncodedDataResults = openpyxl.load_workbook(WIDETABLEFORM_ISO_8859_1_ENCODE_RESULTS_FILENAME)
+        self.WideData_QualifierResults = openpyxl.load_workbook(WIDETABLEFORM_QUALIFIER_RESULTS_FILENAME)
         self.LargeWideDataResults = openpyxl.load_workbook(LARGE_WIDETABLEFORM_RESULTS_FILENAME)
         self.CompoundDataResults = openpyxl.load_workbook(COMPOUNDTABLEFORM_RESULTS_FILENAME)
+        self.CompoundData_QualifierResults = openpyxl.load_workbook(COMPOUNDTABLEFORM_QUALIFIER_RESULTS_FILENAME)
         self.SciexDataResults = openpyxl.load_workbook(SCIEX_RESULTS_FILENAME)
 
     def test_WideData(self):
@@ -52,6 +62,15 @@ class Agilent_Test(unittest.TestCase):
         self.__compare_tables("RT",self.WideData,self.WideDataTransposeResults,transpose=True)
         self.__compare_tables("FWHM",self.WideData,self.WideDataResults)
         self.__compare_tables("FWHM",self.WideData,self.WideDataTransposeResults,transpose=True)
+
+    def test_WideData_Qualifier(self):
+        """Check if the software is able to do the following with WideTableForm_Qualifier.csv:
+
+        * Extract Area successfully using AgilentMSRawData.get_table
+        """
+
+        self.assertEqual("WideTableForm",self.WideData_Qualifier.DataForm)
+        self.__compare_tables("Area",self.WideData_Qualifier,self.WideData_QualifierResults)
 
     def test_ISOEncodedData(self):
         """Check if the software is able to do the following with ISO_8859_1_Encoding.csv:
@@ -83,6 +102,15 @@ class Agilent_Test(unittest.TestCase):
         self.assertEqual("CompoundTableForm",self.CompoundData.DataForm)
         self.__compare_tables("Area",self.CompoundData,self.CompoundDataResults)
         self.__compare_tables("RT",self.CompoundData,self.CompoundDataResults)
+
+    def test_CompoundData_Qualifier(self):
+        """Check if the software is able to do the following with CompoundTableForm_Qualifier.csv:
+
+        * Extract Area successfully using AgilentMSRawData.get_table
+        """
+
+        self.assertEqual("CompoundTableForm",self.CompoundData_Qualifier.DataForm)
+        self.__compare_tables("Area",self.CompoundData_Qualifier,self.CompoundData_QualifierResults)
 
     def test_SciexData(self):
         """Check if the software is able to do the following with large dataset Mohammed_SciEx_data.txt: